+++ OgreMain/src/OgreSubMesh2.cpp
@@ -529,10 +529,10 @@ namespace Ogre {
 
                     uint16 *dstData16 = reinterpret_cast<uint16*>(dstData + acumOffset);
 
-                    dstData16[0] = Math::Clamp( qTangent.x * 32767.0f, -32768.0f, 32767.0f );
-                    dstData16[1] = Math::Clamp( qTangent.y * 32767.0f, -32768.0f, 32767.0f );
-                    dstData16[2] = Math::Clamp( qTangent.z * 32767.0f, -32768.0f, 32767.0f );
-                    dstData16[3] = Math::Clamp( qTangent.w * 32767.0f, -32768.0f, 32767.0f );
+                    dstData16[0] = (uint16)Math::Clamp( qTangent.x * 32767.0f, -32768.0f, 32767.0f );
+                    dstData16[1] = (uint16)Math::Clamp( qTangent.y * 32767.0f, -32768.0f, 32767.0f );
+                    dstData16[2] = (uint16)Math::Clamp( qTangent.z * 32767.0f, -32768.0f, 32767.0f );
+                    dstData16[3] = (uint16)Math::Clamp( qTangent.w * 32767.0f, -32768.0f, 32767.0f );
                 }
                 else if( v1::VertexElement::getBaseType( vElement.mType ) == VET_HALF2 &&
                          v1::VertexElement::getBaseType( itSrc->getType() ) == VET_FLOAT1 )
